generator: Add support for generating the Event Publisher
The generated Event Publisher should only be used in limited scenarios
where only one subscriber is involved and the publisher depends on
the interface provided by the subscriber.

In typical use cases of the publish-subscribe pattern - many subscribers are
involved and the publisher knows nothing about the subscribers - you should
just send the event in the way it should be.